clippy

+++ src/color.rs
@@ -75,8 +75,8 @@ pub fn coloring_from_file(text: String, params: Option<setting::Params>) {
                         } else {
                             let substr = cap.get(0).unwrap().as_str();
                             let len = substr.len();
-                            let color = params.syntaxes[index as usize].color();
-                            comment_now = params.syntaxes[index as usize].ignore_whitespace();
+                            let color = params.syntaxes[index].color();
+                            comment_now = params.syntaxes[index].ignore_whitespace();
                             line_str.push_str(&increasing_str[..increasing_str.len()-len]);
                             line_str.push_str(color);
                             line_str.push_str(substr);
@@ -164,8 +164,8 @@ pub fn coloring_words(serial_buf: &str,
                     } else {
                         let substr = cap.get(0).unwrap().as_str();
                         let len = substr.len();
-                        let color = params.syntaxes[index as usize].color();
-                        *comment_now = params.syntaxes[index as usize].ignore_whitespace();
+                        let color = params.syntaxes[index].color();
+                        *comment_now = params.syntaxes[index].ignore_whitespace();
                         let _ = write!(line_str, "{:\x08<1$}", "", len-1);
                         line_str.push_str(color);
                         line_str.push_str(substr);
